receiver/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
deepgram-sdk==3.2.0
groq==0.4.2
python-dotenv==1.0.0
pyaudio==0.2.14
python-multipart==0.0.6
